+++ mwcp/parsers/ISO.py
@@ -2,7 +2,9 @@
 ISO Image
 """
 
-import isoparser
+from io import BytesIO
+
+import pycdlib
 
 from mwcp import Parser, FileObject
 
@@ -21,27 +23,61 @@ class ImageFile(Parser):
                 return True
         return False
 
-    def recursive_extract(self, record):
+    def walk_handler(self, iso, **kwargs):
         """
-        Recursively extract files within each directory of an ISO image.
+        Uses the iso.walk function to walk the ISO image depending on the path
+        provided in kwargs.
 
-        :param record: A directory within an iso.ISO object
+        :param iso: an open PyCdlib instance
+        :param kwargs: Keyword arguments to be passed to the iso.walk function.
+        Note that this must include either iso_path, udf_path, joliet_path, or
+        rr_path.
+        :return:
         """
-        for child in record.children:
-            if child.is_directory:
-                self.recursive_extract(child)
-            else:
-                self.dispatcher.add(FileObject(child.content, file_name=child.name))
+        for dirname, dirlist, filelist in iso.walk(**kwargs):
+            for filename in filelist:
+                if dirname == "/":
+                    path = dirname + filename
+                else:
+                    path = dirname + "/" + filename
+
+                filedata = BytesIO()
+
+                if "iso_path" in kwargs:
+                    iso.get_file_from_iso_fp(filedata, iso_path=path)
+                elif "udf_path" in kwargs:
+                    iso.get_file_from_iso_fp(filedata, udf_path=path)
+                elif "joliet_path" in kwargs:
+                    iso.get_file_from_iso_fp(filedata, joliet_path=path)
+                elif "rr_path" in kwargs:
+                    iso.get_file_from_iso_fp(filedata, rr_path=path)
+                else:
+                    return
+
+                self.dispatcher.add(FileObject(file_data=filedata.getvalue(), file_name=filename))
 
     def run(self):
         """
-        Create an iso.ISO object and proceed to extract and output embedded files.
+        Walk the ISO image to extract embedded files.
 
         :return:
         """
-        with self.file_object.temp_path() as file_path:
-            try:
-                with isoparser.parse(file_path) as iso:
-                    self.recursive_extract(iso.root)
-            except AttributeError as e:
-                self.logger.warning(f"Error parsing ISO image file: {e}")
+        iso = pycdlib.PyCdlib()
+        iso.open_fp(BytesIO(self.file_object.data))
+
+        try:
+            self.walk_handler(iso, iso_path="/")
+
+            if iso.has_udf():
+                self.logger.info("UDF extension identified")
+                self.walk_handler(iso, udf_path="/")
+
+            if iso.has_joliet():
+                self.logger.info("Joliet extension identified")
+                self.walk_handler(iso, joliet_path="/")
+
+            if iso.has_rock_ridge():
+                self.logger.info("Rock Ridge extension identified")
+                self.walk_handler(iso, rr_path="/")
+        finally:
+            iso.close()